refactor(RestaurantListSection): 레스토랑 리스트 및 아이템 폴더 경로 수정
